Habit Gallery View, Overall UI Theme Change
Changed the habits in the main view to be HStacks
within a scroll view. Trying to change overall
theme of app to a darkish blue almost black.

+++ Project7-9Milestone/HabitTracker/HabitTracker/Habit.swift
@@ -10,12 +10,12 @@ import SwiftUI
 struct Habit: Identifiable, Hashable {
     var id = UUID().uuidString
     var name: String
-    var goal: Int
-    var progress: Int
+    var goal: Double
+    var progress: Double
     var color: Color
     var description: String
     
-    init(name: String, goal: Int, progress: Int, color: Color, description: String) {
+    init(name: String, goal: Double, progress: Double, color: Color, description: String) {
         self.name = name
         self.goal = goal
         self.progress = progress
